Renamed files and folders to plural names.
Moved common.grid_bfs to test_map.py.

Renamed extensions.py to compile.py.

Switched to PyCharm's pre-generated .gitignore.
