+++ src/hades/constants/game_objects.py
