+++ src/hades/game_objects/enemies.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING
 import arcade
 
 # Custom
-from hades.constants.game_object import (
+from hades.constants.game_objects import (
     ARMOUR_INDICATOR_BAR_COLOR,
     ENEMY_INDICATOR_BAR_OFFSET,
     FACING_LEFT,
@@ -21,12 +21,12 @@ from hades.constants.game_object import (
     EntityAttributeType,
     ObjectID,
 )
-from hades.game_objects.attribute import EntityAttribute
+from hades.game_objects.attributes import EntityAttribute
 from hades.game_objects.base import Entity, IndicatorBar
-from hades.game_objects.movement import EnemyMovementManager
+from hades.game_objects.movements import EnemyMovementManager
 
 if TYPE_CHECKING:
-    from hades.constants.game_object import BaseData, EnemyData
+    from hades.constants.game_objects import BaseData, EnemyData
     from hades.views.game_view import Game
 
 __all__ = ("Enemy",)
