+++ src/hades/melee_shader.py
@@ -8,13 +8,13 @@ import struct
 from typing import TYPE_CHECKING
 
 # Custom
-from hades.constants.game_object import MELEE_RESOLUTION, SPRITE_SIZE
+from hades.constants.game_objects import MELEE_RESOLUTION, SPRITE_SIZE
 
 if TYPE_CHECKING:
     from arcade import ArcadeContext
     from arcade.gl import Buffer, Framebuffer, Program, Query
 
-    from hades.game_objects.enemy import Enemy
+    from hades.game_objects.enemies import Enemy
     from hades.views.game_view import Game
 
 __all__ = ("MeleeShader",)
